// src/Models/Clients.ts
export interface Clients{
    idClient: string;
    nameClient: string;
    cpf: string;
    phone:string;
    email:string;
}

export interface ClientsDTO{
    nameClient: string;
    cpf: string;
    phone:string;
    email:string;
}

export interface UpdateClientsDTO{
    idClient: string;
    nameClient: string;
    cpf: string;
    phone:string;
    email:string;
}